requirements.txt
streamlit
tensorflow
numpy
Pillow
opencv-python
matplotlib